Move inspector to proper place.
